requirements.txt
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
beautifulsoup4==4.11.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.7
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
kombu==5.2.4
news==1.0
prompt-toolkit==3.0.38
psycopg2==2.9.5
pyaes==1.6.1
pyasn1==0.4.8
python-crontab==2.7.1
python-dateutil==2.8.2
python-telegram-bot==20.1
pytz==2022.7.1
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
schedule==1.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.3
Telethon==1.27.0
tornado==6.2
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
